move scroll to top button into component, adjust vertical offset observable
