+++ src/app/ranking/ranking-tool/ranking-tool.component.html
@@ -22,7 +22,6 @@
       [areaType]="areaType"
       [dataProperty]="dataProperty"
       [selectedIndex]="selectedIndex"
-      (goToTop)="scrollToTop()"
     ></app-evictions>
     <app-evictors
       class="ranking-component"
